# docs/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        example: ""
        type: string
      message:
        example: Success
        type: string
      status:
        example: 200
        type: integer
    type: object
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      created_at:
        type: string
      icon:
        example: icon-dashboard
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      order_index:
        example: 0
        type: integer
      parent_id:
        type: integer
      path:
        example: /dashboard
        type: string
      title:
        example: Dashboard
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API for hierarchical menu management with reordering and moving
    capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: STK Test API - Menu Management
  version: "1.0"
paths:
  /api/menus:
    get:
      consumes:
      - application/json
      description: Get all menu items in hierarchical tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get all menu items
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: Create a new menu item
      parameters:
      - description: Menu object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create new menu item
      tags:
      - Menus
  /api/menus/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a menu item and its children
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete menu item
      tags:
      - Menus
    get:
      consumes:
      - application/json
      description: Get a single menu item by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get single menu item
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update a menu item
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update menu item
      tags:
      - Menus
  /api/menus/{id}/move:
    patch:
      consumes:
      - application/json
      description: Move a menu item to a different parent
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Move request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Move menu item to different parent
      tags:
      - Menus
  /api/menus/{id}/reorder:
    patch:
      consumes:
      - application/json
      description: Change the order index of a menu item
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reorder request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Reorder menu item within same level
      tags:
      - Menus
  /health:
    get:
      consumes:
      - application/json
      description: Check API health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health Check
      tags:
      - Health
schemes:
- http
- https
swagger: "2.0"